chore(refactor): move `RequiresOnline` into controllers
