+++ src/pages/invite/Invite.tsx
@@ -12,7 +12,6 @@ import { TextReact } from "../../lib/i18n";
 
 import { useApplicationState } from "../../mobx/State";
 
-import RequiresOnline from "../../context/revoltjs/RequiresOnline";
 import { takeError } from "../../context/revoltjs/util";
 
 import ServerIcon from "../../components/common/ServerIcon";
@@ -21,6 +20,7 @@ import {
     useClient,
     useSession,
 } from "../../controllers/client/ClientController";
+import RequiresOnline from "../../controllers/client/jsx/RequiresOnline";
 
 export default function Invite() {
     const history = useHistory();
